fixing codacy issues
