+++ src/main/java/com/butbetter/applicationservices/ApplicationServices.java
@@ -1,4 +1,4 @@
-package com.butbetter.applicationServices;
+package com.butbetter.applicationservices;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
