+++ src/test/java/com/butbetter/applicationservices/ApplicationServicesApplicationTests.java
@@ -1,4 +1,4 @@
-package com.butbetter.applicationServices;
+package com.butbetter.applicationservices;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
